Add command to client
